+++ apps/native-component-list/src/navigation/ExpoApis.ts
@@ -18,6 +18,7 @@ const AsyncStorage = optionalRequire(() => require('../screens/AsyncStorageScree
 const Audio = optionalRequire(() => require('../screens/AV/AudioScreen'));
 const AuthSession = optionalRequire(() => require('../screens/AuthSession/AuthSessionScreen'));
 const BackgroundFetch = optionalRequire(() => require('../screens/BackgroundFetchScreen'));
+const BackgroundLocation = optionalRequire(() => require('../screens/BackgroundLocationMapScreen'));
 const Battery = optionalRequire(() => require('../screens/BatteryScreen'));
 const Branch = optionalRequire(() => require('../screens/BranchScreen'));
 const Brightness = optionalRequire(() => require('../screens/BrightnessScreen'));
@@ -92,6 +93,7 @@ const optionalScreens: {
   AsyncStorage,
   Audio,
   AuthSession,
+  BackgroundLocation,
   BackgroundFetch,
   Battery,
   Branch,