+++ ios/Exponent/Kernel/AppLoader/EXAppLoaderExpoUpdates.m
@@ -38,6 +38,13 @@ NS_ASSUME_NONNULL_BEGIN
 @property (nonatomic, assign) BOOL shouldShowRemoteUpdateStatus;
 @property (nonatomic, assign) BOOL isUpToDate;
 
+/**
+ * Stateful variable to let us prevent multiple simultaneous fetches from the development server.
+ * This can happen when reloading a bundle with remote debugging enabled;
+ * RN requests the bundle multiple times for some reason.
+ */
+@property (nonatomic, assign) BOOL isLoadingDevelopmentJavaScriptResource;
+
 @property (nonatomic, strong, nullable) NSError *error;
 
 @property (nonatomic, assign) BOOL shouldUseCacheOnly;
@@ -89,6 +96,7 @@ NS_ASSUME_NONNULL_BEGIN
   _remoteUpdateStatus = kEXAppLoaderRemoteUpdateStatusChecking;
   _shouldShowRemoteUpdateStatus = YES;
   _isUpToDate = NO;
+  _isLoadingDevelopmentJavaScriptResource = NO;
 }
 
 - (EXAppLoaderStatus)status
@@ -130,6 +138,13 @@ NS_ASSUME_NONNULL_BEGIN
                                  userInfo:@{}];
   }
   NSAssert([self supportsBundleReload], @"Tried to force a bundle reload on a non-development bundle");
+  if (self.isLoadingDevelopmentJavaScriptResource) {
+    // prevent multiple simultaneous fetches from the development server.
+    // this can happen when reloading a bundle with remote debugging enabled;
+    // RN requests the bundle multiple times for some reason.
+    // TODO: fix inside of RN
+    return;
+  }
   [self _loadDevelopmentJavaScriptResource];
 }
 
@@ -410,6 +425,7 @@ NS_ASSUME_NONNULL_BEGIN
 
 - (void)_loadDevelopmentJavaScriptResource
 {
+  _isLoadingDevelopmentJavaScriptResource = YES;
   EXAppFetcher *appFetcher = [[EXAppFetcher alloc] initWithAppLoader:self];
   [appFetcher fetchJSBundleWithManifest:self.optimisticManifest cacheBehavior:EXCachedResourceNoCache timeoutInterval:kEXJSBundleTimeout progress:^(EXLoadingProgress *progress) {
     if (self.delegate) {
@@ -418,11 +434,13 @@ NS_ASSUME_NONNULL_BEGIN
   } success:^(NSData *bundle) {
     self.isUpToDate = YES;
     self.bundle = bundle;
+    self.isLoadingDevelopmentJavaScriptResource = NO;
     if (self.delegate) {
       [self.delegate appLoader:self didFinishLoadingManifest:self.optimisticManifest bundle:self.bundle];
     }
   } error:^(NSError *error) {
     self.error = error;
+    self.isLoadingDevelopmentJavaScriptResource = NO;
     if (self.delegate) {
       [self.delegate appLoader:self didFailWithError:error];
     }