+++ packages/expo-ads-facebook/src/AdSettings.ts
@@ -50,6 +50,21 @@ export default {
   },
 
   /**
+   * Sets whether Facebook SDK should enable advertising tracking.
+   */
+  setAdvertiserTrackingEnabled(enabled: boolean): void {
+    // noop outside of iOS
+    if (Platform.OS !== 'ios') {
+      return;
+    }
+
+    if (!CTKAdSettingsManager.setAdvertiserTrackingEnabled) {
+      throw new UnavailabilityError('expo-ads-facebook', 'setAdvertiserTrackingEnabled');
+    }
+    CTKAdSettingsManager.setAdvertiserTrackingEnabled(enabled);
+  },
+
+  /**
    * Registers given device with `deviceHash` to receive test Facebook ads.
    */
   addTestDevice(deviceHash: string): void {