+++ tools/src/prebuilds/Prebuilder.ts
@@ -78,6 +78,7 @@ export const PACKAGES_TO_PREBUILD = [
   'expo-splash-screen',
   // 'expo-sqlite',
   // 'expo-store-review',
+  'expo-structured-headers',
   // 'expo-task-manager',
   'expo-updates',
   // 'expo-video-thumbnails',