+++ packages/expo-apple-authentication/CHANGELOG.md
@@ -6,6 +6,12 @@
 
 ### 🎉 New features
 
+### 🐛 Bug fixes
+
+## 3.1.0 — 2021-03-10
+
+### 🎉 New features
+
 - Created config plugin ([#11979](https://github.com/expo/expo/pull/11979) by [@EvanBacon](https://github.com/EvanBacon))
 
 ### 🐛 Bug fixes