+++ packages/expo-local-authentication/CHANGELOG.md
@@ -4,6 +4,14 @@
 
 ### 🛠 Breaking changes
 
+### 🎉 New features
+
+### 🐛 Bug fixes
+
+## 11.0.0 — 2021-03-10
+
+### 🛠 Breaking changes
+
 - Remove deprecated support for passing a string to `authenticateAsync` in favor of the `promptMessage` option. ([#11906](https://github.com/expo/expo/pull/11906) by [@EvanBacon](https://github.com/EvanBacon))
 
 ### 🎉 New features