+++ packages/unimodules-font-interface/CHANGELOG.md
@@ -6,9 +6,13 @@
 
 ### 🎉 New features
 
-- Updated Android build configuration to target Android 11 (added support for Android SDK 30). ([#11647](https://github.com/expo/expo/pull/11647) by [@bbarthec](https://github.com/bbarthec))
-
 ### 🐛 Bug fixes
+
+## 6.1.0 — 2021-03-10
+
+### 🎉 New features
+
+- Updated Android build configuration to target Android 11 (added support for Android SDK 30). ([#11647](https://github.com/expo/expo/pull/11647) by [@bbarthec](https://github.com/bbarthec))
 
 ## 6.0.0 — 2021-01-15